+++ packages/common/web/lib/normalization.js
@@ -1,1 +1,1 @@
-export { default as normalizeInitialState } from '../src/Utils/normalization/normalizeInitialState';
+export { normalizeInitialState } from 'ricos-content';