+++ packages/common/web/src/index.ts
@@ -13,18 +13,62 @@ export {
   textWrapClassName,
   depthClassName,
 } from './Utils/classNameStrategies';
-export {
-  isValidExactUrl,
-  isValidUrl,
-  normalizeUrl,
-  getUrlMatches,
-  startsWithHttps,
-  hasProtocol,
-} from './Utils/urlValidators';
 export { mergeStyles } from './Utils/mergeStyles';
-export { default as normalizeInitialState } from './Utils/normalization/normalizeInitialState';
+
+// export {
+//   normalizeInitialState,
+//   isValidExactUrl,
+//   isValidUrl,
+//   normalizeUrl,
+//   getUrlMatches,
+//   startsWithHttps,
+//   hasProtocol,
+//   RicosInlineStyleRange,
+//   RicosEntityRange,
+//   RicosContentBlock,
+//   RicosEntity,
+//   RicosContent,
+//   HEADER_BLOCK,
+//   BLOCK_TYPES,
+//   UNSTYLED,
+//   BLOCKQUOTE,
+//   LINK_BUTTON_TYPE,
+//   ACTION_BUTTON_TYPE,
+//   CODE_BLOCK_TYPE,
+//   DIVIDER_TYPE,
+//   EMOJI_TYPE,
+//   FILE_UPLOAD_TYPE,
+//   GALLERY_TYPE,
+//   GIPHY_TYPE,
+//   HASHTAG_TYPE,
+//   HEADERS_MARKDOWN_TYPE,
+//   HTML_TYPE,
+//   IMAGE_TYPE,
+//   IMAGE_TYPE_LEGACY,
+//   INDENT_TYPE,
+//   LINE_SPACING_TYPE,
+//   HEADINGS_DROPDOWN_TYPE,
+//   SPOILER_TYPE,
+//   EXTERNAL_LINK_TYPE,
+//   LINK_TYPE,
+//   LINK_PREVIEW_TYPE,
+//   MAP_TYPE,
+//   EXTERNAL_MENTIONS_TYPE,
+//   MENTION_TYPE,
+//   SOUND_CLOUD_TYPE,
+//   TEXT_COLOR_TYPE,
+//   TEXT_HIGHLIGHT_TYPE,
+//   UNDO_REDO_TYPE,
+//   VERTICAL_EMBED_TYPE,
+//   VIDEO_TYPE,
+//   VIDEO_TYPE_LEGACY,
+//   POLL_TYPE,
+//   Version,
+// } from 'ricos-content';
+export * from 'ricos-content';
+
 export { default as getDisplayName } from './Utils/getDisplayName';
-export { default as Version } from './Utils/versioningUtils';
+
 export { hasLinksInBlock, getLinkRangesInBlock } from './Utils/draftUtils';
 export { validate, getContentStateSchema } from './Utils/data-schema-validator';
 export { isSSR } from './Utils/ssrUtils';
@@ -35,4 +79,3 @@ export { GlobalContext } from './Utils/contexts';
 export { getImageSrc, WIX_MEDIA_DEFAULT } from './Utils/imageUtils';
 export { isHexColor } from './Utils/colorUtils';
 export { isRtl, getLangDir } from './Utils/rtlUtils';
-export * from './consts';