+++ packages/common/web/src/types/commonTypes.ts
@@ -1,6 +1,6 @@
 import { CSSProperties } from 'react';
 import { Styles as ReactModalStyles } from 'react-modal';
-
+import { ComponentData } from '../index';
 export type ModalStyles = ReactModalStyles;
 export type Styles = Record<string, CSSProperties>;
 
@@ -21,20 +21,5 @@ export type ClassNameStrategy = (
 
 export type ContainerClassNameStrategy = (theme: RichContentTheme) => CSSProperties;
 
-export interface ComponentData {
-  config?: { alignment?: string; size?: string; url?: string; textWrap?: string };
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  src?: any;
-  srcType?: string;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [propName: string]: any;
-}
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type TranslateFunction = (key: string, template?: any) => string;
-
-export type NormalizeConfig = {
-  anchorTarget?: string;
-  relValue?: string;
-  disableInlineImages?: boolean;
-  removeInvalidInlinePlugins?: boolean;
-};