+++ packages/common/web/src/types/index.ts
@@ -1,7 +1,8 @@
+export { ComponentData, RicosContentBlock, RicosContent, RicosEntityRange } from 'ricos-content';
 export * from './buttonTypes';
 export * from './themeTypes';
 export * from './commonTypes';
-export * from './contentTypes';
+
 export * from './editorTypes';
 export * from './mediaUploadTypes';
 export * from './pluginTypes';