+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -1,8 +1,13 @@
-import { Decorator, Helpers, GetToolbarSettings, PluginTypeMapper } from 'wix-rich-content-common';
+import {
+  Decorator,
+  Helpers,
+  GetToolbarSettings,
+  PluginTypeMapper,
+  RicosContent,
+} from 'wix-rich-content-common';
 import { EditorState, EditorProps } from 'draft-js';
 import { ReactElement } from 'react';
 import {
-  RicosContent,
   RicosCssOverride,
   InlineStyleMapper,
   ModalsMap,