+++ packages/ricos-common/web/src/index.ts
@@ -3,3 +3,4 @@ export { RicosEngine } from './RicosEngine';
 export { shouldRenderChild, emptyState } from './utils';
 export { default as localeStrategy } from './localeStrategy/localeStrategy';
 export { DRAFT_EDITOR_PROPS } from './consts';
+export { RicosContent } from 'wix-rich-content-common';