+++ packages/ricos-common/web/src/types.ts
@@ -2,4 +2,3 @@ export * from './previewStrategy/previewTypes';
 export * from './pluginsStrategy/pluginTypes';
 export * from './themeTypes';
 export * from './RicosTypes';
-export { RicosContent } from 'wix-rich-content-common';