+++ CHANGELOG.md
@@ -21,15 +21,20 @@
   - [#1542](https://github.com/wix-incubator/rich-content/pull/1542) fixed appearance of inlineToolbarButtons to the new design
   - [#1548](https://github.com/wix-incubator/rich-content/pull/1548) accept only valid draft-js props in `draftEditorSettings`
 
+### :rocket: New Feature
+- `ricos-content`
+  - [#1527](https://github.com/wix-incubator/rich-content/pull/1527) Platform agnostic package for handling Ricos content 
+
 ### 🏠 Internal
 - `general`
-  - [#1544](https://github.com/wix-incubator/rich-content/pull/15441) one-line index files removed
+  - [#1544](https://github.com/wix-incubator/rich-content/pull/1544) one-line index files removed
 - `exampleApp`
   - [#1541](https://github.com/wix-incubator/rich-content/pull/1541) file upload native\media manager toggle in gear icon
   - [#1537](https://github.com/wix-incubator/rich-content/pull/1537) bi callback `onViewerAction` outputs to console.log
 
 ### :book: Documentation
 - [#1546](https://github.com/wix-incubator/rich-content/pull/1546) external toolbar customization docs
+
 
 </details>
 <hr/>